skills/systematic-literature-review/requirements.txt
pyyaml>=6.0
requests>=2.31
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
